SOLID : LSP 활용하기
